backend/pyproject.toml
[tool.poetry]
name = "macros-backend"
version = "0.1.0"
description = ""
authors = ["Ryan <rjl275@cornell.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
flask = "^2.3.3"
neo4j = "^5.12.0"
flask-restful = "^0.3.10"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
